Minimalist fix to account for GNU C dialect version, and better names for HAVE_ float feature macros
Some behaviour of CIL depends on the target machine/ABI, the target
compiler, but also its version and any options that may be in effect.
This is all a big mess at present; here we add a necessary notion of GNU
C version, to allow us to tokenize floating-point names that are only
present from gcc 7.0 onwards but can be compatibility-typedef'd by
header files included by earlier version. Most importantly, add a
comment explaining how this needs to be cleaned up.

Also, make the names for floating-point feature macros more descriptive,
as noted in 7c550e93.