Support `[[fallthrough]]` statements
There's a new statement from C17 --- `[[fallthrough]]` --- which we
don't support yet. GNUTLS uses it.

It seems a little redundant to me? Seemingly it's supposed to indicate
to the compiler that fallthrough really is expected somewhere. This adds
it as a new type of statement, albeit with some hacky bits in the lexer.